backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
boto3>=1.28.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
python-dotenv>=1.0.0
requests>=2.28.0
numpy>=1.21.0
python-socketio>=5.8.0
eventlet>=0.33.0
